+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+0.2
+---
+
+* [BC BREAK] Remove `Agent` and `MultiAgent` suffixes from injection aliases.
+* Add `http_client` option to VertexAI platform configuration
+* Add `bedrock` support for platform configuration
+
 0.1
 ---